TW-403: refactor chat_list_item into smaller widgets
